third_party/WebKit/LayoutTests/fast/notifications/notification-properties.html

<!doctype html>
<html>
    <head>
        <script src="../../resources/js-test.js"></script>
    </head>
    <body>
        <div id="description"></div>
        <div id="console"></div>
        <script>
            description("Checks that the {title, dir, lang, body, tag, icon} properties are set on the Notification object.");
            if (window.testRunner) {
                testRunner.grantWebNotificationPermission("file://", true);
                testRunner.dumpAsText();
            }

            var options = {
                dir: "rtl",
                lang: "nl-NL",
                body: "Hallo, wereld!",
                tag: "notification",
                icon: "http://localhost/my_icon.png"
            };

            var notification = new Notification("My Notification", options);

            shouldBeEqualToString("notification.title", "My Notification");
            shouldBeEqualToString("notification.dir", options.dir);
            shouldBeEqualToString("notification.lang", options.lang);
            shouldBeEqualToString("notification.body", options.body);
            shouldBeEqualToString("notification.tag", options.tag);
            shouldBeEqualToString("notification.icon", options.icon);

            var emptyNotification = new Notification("My Notification");

            shouldBeEqualToString("emptyNotification.title", "My Notification");
            shouldBeEqualToString("emptyNotification.dir", "auto");
            shouldBeEqualToString("emptyNotification.lang", "");
            shouldBeEqualToString("emptyNotification.body", "");
            shouldBeEqualToString("emptyNotification.tag", "");
            shouldBeEqualToString("emptyNotification.icon", "");

            var invalidIconNotification = new Notification("My Notification", {
                icon: "http://test:test/"
            });

            // Invalid icon URLs should be reset to an empty string.
            shouldBeEqualToString("invalidIconNotification.icon", "");

            var serializedUrlNotification = new Notification("My Notification", {
                icon: "http://example.com"
            });

            // Icon URLs should be returned in serialized form.
            shouldBeEqualToString("serializedUrlNotification.icon", "http://example.com/");

            var noTagNotification = new Notification("My Notification"),
                emptyTagNotification = new Notification("My Notification", { tag: "" });

            // Setting an empty string as the tag should be equal to not setting the tag at all.
            shouldBe("noTagNotification.tag", "emptyTagNotification.tag");

            var successfullyParsed = true;
        </script>
    </body>
</html>
